VDB-5145 removing windows compiler warnings from klib (1st wave)

+++ libs/klib/namelist.c
@@ -163,7 +163,7 @@ LIB_EXPORT bool KNamelistContains( const KNamelist * self, const char * to_find 
             if ( 0 == rc ) {
                 size_t entry_size = string_size( entry );
                 if ( to_find_size == entry_size ) {
-                    int cmp = string_cmp( to_find, to_find_size, entry, entry_size, entry_size );
+                    int cmp = string_cmp( to_find, to_find_size, entry, entry_size, (uint32_t)entry_size );
                     res = ( 0 == cmp );
                 }
             }
