+++ libs/klib/pbstree-impl.c
@@ -419,7 +419,7 @@ void CC PBSTreeImplForEach16 ( const PBSTree *self, bool reverse,
 	if ( ! reverse )
 	  {
 	    GET16 ( off, pt -> data_idx . v16 [ 0 ] );
-	    for ( id = 1; id < num_nodes; off = end, ++ id )
+	    for ( id = 1; id < num_nodes; off = (uint16_t)end, ++ id )
 	      {
 		uint16_t end16;
 		GET16 ( end16, pt -> data_idx . v16 [ id ] );
@@ -600,7 +600,7 @@ bool CC PBSTreeImplDoUntil16 ( const PBSTree *self, bool reverse,
         if ( ! reverse )
         {
             GET16 ( off, pt -> data_idx . v16 [ 0 ] );
-            for ( id = 1; id < num_nodes; off = end, ++ id )
+            for ( id = 1; id < num_nodes; off = (uint16_t)end, ++ id )
             {
                 uint16_t end16;
                 GET16 ( end16, pt -> data_idx . v16 [ id ] );
