+++ libs/klib/zz_json-lex.c
@@ -2780,7 +2780,7 @@ void
 JsonScan_yylex_init ( JsonScanBlock* sb, const char *str, size_t len )
 {
     yylex_init ( & sb -> scanner );
-    sb -> buffer = yy_scan_bytes ( ( yyconst char * ) str, len, sb -> scanner );
+    sb -> buffer = yy_scan_bytes ( ( yyconst char * ) str, (int) len, sb -> scanner );
     ( ( struct yyguts_t * ) sb -> scanner ) -> yyextra_r = sb; /* back pointer to the scan block */
     sb -> error = NULL;
 }
